/apps/[slug]/setup
