+++ apps/web/app/future/apps/[slug]/setup/page.tsx
@@ -1,4 +1,3 @@
-import Page from "@pages/apps/[slug]/setup";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { _generateMetadata } from "app/_utils";
 import { WithLayout } from "app/layoutHOC";
@@ -6,6 +5,8 @@ import type { InferGetServerSidePropsType } from "next";
 
 import { getServerSideProps } from "@calcom/app-store/_pages/setup/_getServerSideProps";
 import { APP_NAME } from "@calcom/lib/constants";
+
+import SetupInformation from "@components/pages/apps/[slug]/setup";
 
 export const generateMetadata = async ({ params }: { params: Record<string, string | string[]> }) => {
   return await _generateMetadata(
@@ -18,4 +19,4 @@ type T = InferGetServerSidePropsType<typeof getServerSideProps>;
 
 const getData = withAppDirSsr<T>(getServerSideProps);
 
-export default WithLayout({ getLayout: null, Page, getData });
+export default WithLayout({ getLayout: null, Page: SetupInformation, getData });
